+++ modules/notationengine.js
@@ -1,66 +1,33 @@
 function timeNotation(ms) {
-    let time = ms.toString();
-    if (ms < 1000) {
-        return(`${time[0] + time[1] + time[2]}ms`);
-    }
+    const dateObj = new Date(ms);
+    const hours = dateObj.getUTCHours();
+    const minutes = dateObj.getUTCMinutes();
+    const seconds = dateObj.getSeconds();
+    const milliseconds = ms.toString().slice(-3);
+    const timeString = hours.toString().padStart(2, '0') + ':' +  minutes.toString().padStart(2, '0') + ':' + seconds.toString().padStart(2, '0') + "." + milliseconds;
 
-    if (ms < 10_000) {
-        return(`${time[0]}.${time[1]}s`);
-    }
-
-    if (ms < 100_000) {
-        return(`${time[0] + time[1]}.${time[2]}`);
-    }
+    return timeString;
 }
 
 function biNotation(vala, decpointa, notationa) {
     let notation = notationa ?? 0;
     let val = new Decimal(vala);
     let exp = new Decimal(floorLog10(val));
-    let decpoint = new Decimal(decpointa);
+    let decpoint = new Decimal(decpointa) ?? 3;
     let valstring = val.toFixed(decpoint);
 
 
     if (val.lt(1000)) {
-        if (val.equals(0)) {
-            return("0");
-        }
-        return(valstring);
+        return(val.equals(0) ? "0" : valstring);
     }
 
     if (notation == 0) {
-        if (exp.lt(33)) { //1 decillion
-            return(StandardNotation(valstring, exp));
-        } else {
-            return(SciNotation(valstring, exp));
-        }
-    } else if (notation == 1) {
-        return(StandardNotation(valstring, exp));
-    } else if (notation == 2) {
-        return(SciNotation(valstring, exp));
+        if (exp.lt(33)) return(StandardNotation(valstring, exp));
+        else return(SciNotation(valstring, exp));
     }
-}
-
-function notation(val, decpoint, notation) {
-    let exp = Math.floor(Math.log10(val));
-    let valstring = val.toString();
-    if (notation = null) { notation = 0; }
-    if (decpoint = null) { decpoint = 2; }
-
-    if (val < 1000) {
-        return(valstring.toFixed(decpoint));
-    }
+    else if (notation == 1) return(StandardNotation(valstring, exp));
+    else if (notation == 2) return(SciNotation(valstring, exp));
     
-    if (notation == 0) {
-        if (val < 1_000_000_000_000_000_000_000_000_000_000_000) { //1 decillion
-            return(StandardNotation(valstring, exp));
-        }
-        return(SciNotation(valstring, exp));
-    } else if (notation == 1) {
-        return(StandardNotation(valstring, exp));
-    } else if (notation == 2) {
-        return(SciNotation(valstring, exp));
-    }
 }
 
 function floorLog10(val) {
@@ -72,21 +39,15 @@ function floorLog10(val) {
 
 function StandardNotation(valstring, exp) {
     let suffixes = ["", "K", "M", "B", "T", "Qa", "Qt", "Sx", "Sp", "Oc", "No", "Dc"];
-
-    let q;
-    if (exp < 18) { q = 0; } else { q = 1;}
-
+    let q = exp < 18 ? 0 : 1;
 
     switch(exp % 3) {
-        case 0:
-            return(valstring[0] + "." + valstring[1+q] + valstring[2+q] + valstring[3+q] + suffixes[Math.floor(exp/3)]);
-        case 1:
-            return(valstring[0] + valstring[1+q] + "." + valstring[2+q] + valstring[3+q] + valstring[4+q] + suffixes[Math.floor(exp/3)]);
-        case 2:
-            return(valstring[0] + valstring[1+q] + valstring[2+q] + "." + valstring[3+q] + valstring[4+q] + valstring[5+q] + suffixes[Math.floor(exp/3)]);
+        case 0: return(valstring[0] + "." + valstring[1+q] + valstring[2+q] + valstring[3+q] + suffixes[Math.floor(exp/3)]);
+        case 1: return(valstring[0] + valstring[1+q] + "." + valstring[2+q] + valstring[3+q] + valstring[4+q] + suffixes[Math.floor(exp/3)]);
+        case 2: return(valstring[0] + valstring[1+q] + valstring[2+q] + "." + valstring[3+q] + valstring[4+q] + valstring[5+q] + suffixes[Math.floor(exp/3)]);
     }
 }
 
-function SciNotation(valstring, exp) { return(valstring[0] + "." + valstring[2] + valstring[3] + valstring[4] + "e" + exp.toString()); }
+function SciNotation(valstring, exp) { return(valstring[0] + "." + valstring.substring(2, 4) + "e" + biNotation(exp, 0)); }
 
-export {notation, biNotation, floorLog10, timeNotation}; 
+export { biNotation, timeNotation};