+++ changelog.md
@@ -1,4 +1,26 @@
 # Version Log
 
+### Version 0.3-BETA (DRAFT)
+* Added way to unlock production slots
+* Added more element unlocking requirements (to be balanced)
+* Worked on Save/Load system
+  * Added Local save system (browser local storage)
+  * Added Import Save System
+  * Added Autosave System
+* Added Settings Menu
+* Added Loading screen
+* Added Overview Menu
+* Added Offline Tick simulation
+
+### Version 0.2.1-BETA
+* Fixed bug: invalid crafting bond breaks production item
+* changed production item structure
+
+### Version 0.2-BETA
+* Introduced inventory
+* Introduced Unlocking
+* converted element count to Decimal()
+* Added Progress bars to Production
+
 ### Version 0.1-BETA
 * initial release